Some cleanup
Since this repository was created from prototyping there is some
cleanup todo. Expect so less structured cleanup commits coming...
