+++ package.json
@@ -4,16 +4,13 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "npm run build:swagger",
-    "build:docs": "spectacle ./build/swagger.json -t docs",
-    "build:swagger": "node ./src/createSpecifications",
-    "build:swagger:watch": "node ./src/watch",
-    "install": "npm run build:swagger",
+    "build": "node ./src/build",
+    "install": "npm run build",
     "lint:js": "eslint 'src/**/*.js'",
     "pretty": "prettier --single-quote --write 'src/**/*.{js,jsx}'",
-    "start": "npm-run-all -p build:swagger:watch start:spectacle start:node-swagger",
-    "start:node-swagger": "node src/server.js",
-    "start:spectacle": "spectacle -d ./build/swagger.json -t docs",
+    "start": "npm run build && npm run start:openApi",
+    "start:openApi": "docker run -p 80:8080 -e SWAGGER_JSON=/build/openApi.json -v $(pwd)/build:/build swaggerapi/swagger-ui",
+    "start:swagger": "docker run -p 80:8080 -e SWAGGER_JSON=/build/swagger.json -v $(pwd)/build:/build swaggerapi/swagger-ui",
     "test": "npm run lint:js"
   },
   "author": "",
